Fixed bug #12375: typo3temp got filled with thousands of EXTERNAL_URL* files
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@7213 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ typo3/sysext/indexed_search/class.indexer.php
@@ -914,12 +914,14 @@ class tx_indexedsearch_indexer {
 			if (strlen($content))	{
 
 					// Create temporary file:
-				$tmpFile = t3lib_div::tempnam('EXTERNAL_URL').'.html';
-				t3lib_div::writeFile($tmpFile, $content);
+				$tmpFile = t3lib_div::tempnam('EXTERNAL_URL');
+				if ($tmpFile) {
+					t3lib_div::writeFile($tmpFile, $content);
 
-					// Index that file:
-				$this->indexRegularDocument($externalUrl, TRUE, $tmpFile, 'html');	// Using "TRUE" for second parameter to force indexing of external URLs (mtime doesn't make sense, does it?)
-				unlink($tmpFile);
+						// Index that file:
+					$this->indexRegularDocument($externalUrl, TRUE, $tmpFile, 'html');	// Using "TRUE" for second parameter to force indexing of external URLs (mtime doesn't make sense, does it?)
+					unlink($tmpFile);
+				}
 			}
 		}
 	}
